// model/meal/IngredientCategory.ts
export enum IngredientCategory {
  uncategorized = 'uncategorized',
  fruits = 'fruits',
  vegetables = 'vegetables',
  cannedFood = 'cannedFood',
  frozenGoods = 'frozenGoods',
  meat = 'meat',
  fish = 'fish',
  deli = 'deli',
  dairyAndEggs = 'dairyAndEggs',
  condimentsAndSpices = 'condimentsAndSpices',
  saucesAndOil = 'saucesAndOil',
  snacks = 'snacks',
  breadAndBakery = 'breadAndBakery',
  beverages = 'beverages',
  pastaAndRice = 'pastaAndRice',
  cereal = 'cereal',
  bakingSupplies = 'bakingSupplies'
}